+++ apple/WCDB/interface/statictics/WCTStatictics.mm
@@ -46,7 +46,7 @@
                 [dictionary setObject:@(iter.first.c_str())
                                forKey:@(iter.second)];
             }
-            trace(tag, dictionary, cost);
+            trace(tag, dictionary, (NSUInteger)cost);
         });
     }else {
         WCDB::Database::SetGlobalTrace(nullptr);
